feat: adds audio interop + test project
I needed this library for audio, but alas, audio is not a part of this project.

Consequently, I have extended functionality to include audio.

I also put together a bare-bones example to test the audio. I wasn't going to check it in,
but there's no harm in doing so since when building the nuspec library it can simply be ignored.

Todo:
Add more interop, i would like to be able to use IO streams directly so i
don't need to worry with the file system.

Vendor the osx + linux binaries directly (instead of relying on the redist), suffix
the libaries with the architecture and tweak the loader to select the correct dll based on
architecture.

I may start a new repository that provides nice abstractions for the audio component of this library.